(Lib + Networking) split middleware
